% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProportion.R
\name{bootstrap_CI}
\alias{bootstrap_CI}
\title{Calculates the 95\% confidence intervals for each bootstrap iteration}
\usage{
bootstrap_CI(bootstrap_percentage_output)
}
\arguments{
\item{bootstrap_percentage_output}{The dataframe output from bootstrap_percentage function}
}
\value{
A dataframe of 95\% confidence intervals
}
\description{
Calculates the 95\% confidence intervals for each bootstrap iteration
}
