% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProportion.R
\name{bootstrap_percentage}
\alias{bootstrap_percentage}
\title{Bootstraps the samples and recalculates proportion of samples containing a property}
\usage{
bootstrap_percentage(df_map, level = "Drug.Class", B)
}
\arguments{
\item{df_map}{A dataframe of combined non-subsampled or subsampled mapping data and metadata}

\item{level}{A string of a column name e.g. Drug.Class}

\item{B}{A numerical value of the number of types to sample in bootstrapping}
}
\value{
A dataframe of proportions of samples for each bootstrap iteration
}
\description{
Bootstraps the samples and recalculates proportion of samples containing a property
}
