% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARGPhageAnalysis.R
\name{combineARGsAndPhages}
\alias{combineARGsAndPhages}
\title{Combines BLAST output data from CARD and PATRIC}
\usage{
combineARGsAndPhages(phages_data, args_data)
}
\arguments{
\item{phages_data}{A dataframe containing combined phage blast data}

\item{args_data}{A dataframe containing combined ARG blast data}
}
\description{
Combines BLAST output data from CARD and PATRIC
}
\examples{

}
