% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineBedtools.R
\name{combineBedtools}
\alias{combineBedtools}
\title{Combine mapping data and metadata for all samples and write to a csv file}
\usage{
combineBedtools(filenames, output_csv, metadata_csv, card_dir,
  subsampled = FALSE)
}
\arguments{
\item{filenames}{A character vector of filenames of mapping data from Bedtools}

\item{output_csv}{A string of the file path of the output csv file}

\item{metadata_csv}{A string of the file path of the metadata file}

\item{card_dir}{A string of the file directory of CARD metadata}

\item{subsampled}{Logical: if TRUE, the filenames will be split into IDs and the sequence number}
}
\value{
None
}
\description{
Combine mapping data and metadata for all samples and write to a csv file
}
