% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processFiles.R
\name{combineCARDBlast}
\alias{combineCARDBlast}
\title{Combines CARD blast output files with metadata and CARD data}
\usage{
combineCARDBlast(filenames, card_dir, metadata_file, output_file)
}
\arguments{
\item{filenames}{A character vector of bedtools filenames}

\item{card_dir}{A string of the CARD metadata directory that includes aro_index and aro_categories_index}

\item{metadata_file}{A string of the location of the metadata file}

\item{output_file}{A string of the location of the output}
}
\description{
Combines CARD blast output files with metadata and CARD data
}
\examples{

}
