% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineMetaphlanAndARGs.R
\name{combineMetaphlanandARG}
\alias{combineMetaphlanandARG}
\title{Combine a dataframe containing mapping data and metadata data with a dataframe of Metaphlan relative abundances}
\usage{
combineMetaphlanandARG(df_map, metaphlan)
}
\arguments{
\item{df_map}{A dataframe of combined non-subsampled or subsampled mapping data and metadata}

\item{metaphlan}{A dataframe of Metaphlan relative abundances of all samples}
}
\value{
A dataframe combining rpkm from mapping data and relative abundance from metaphlan with row names as IDs
}
\description{
Combine a dataframe containing mapping data and metadata data with a dataframe of Metaphlan relative abundances
}
