% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processFiles.R
\name{combineMetaphlanSamples}
\alias{combineMetaphlanSamples}
\title{Combines metaphlan2 output data}
\usage{
combineMetaphlanSamples(filenames, output_file)
}
\arguments{
\item{filenames}{A character vector of metaphlan2 output filenames}

\item{output_file}{A string of the location of combined metaphlan2 file}
}
\description{
Combines metaphlan2 output data
}
\examples{

}
