% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processFiles.R
\name{combinePATRICBlast}
\alias{combinePATRICBlast}
\title{Combines PATRIC blast output files with metadata and PATRIC data}
\usage{
combinePATRICBlast(filenames, patric_headers, metadata_file, output_file)
}
\arguments{
\item{filenames}{A character vector of bedtools filenames}

\item{patric_headers}{A dataframe of the faa headers in the patric database}

\item{metadata_file}{A string of the location of the metadata file}

\item{output_file}{A string of the location of the output}
}
\description{
Combines PATRIC blast output files with metadata and PATRIC data
}
\examples{

}
