% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProportion.R
\name{combineSubsampled}
\alias{combineSubsampled}
\title{Combines proportions and confidence intervals from non-subsampled and equivalent subsampled samples}
\usage{
combineSubsampled(df_map, df_map_sub, level, sample_type)
}
\arguments{
\item{df_map}{A dataframe of combined non-subsampled mapping data and metadata}

\item{df_map_sub}{A dataframe of combined subsampled mapping data and metadata}

\item{level}{A string of a column name e.g. "Drug.Class"}

\item{sample_type}{A string of a sample type e.g. "saliva"}
}
\value{
A dataframe with proportions and confidence intervals from non-subsampled and equivalent subsampled samples
}
\description{
Combines proportions and confidence intervals from non-subsampled and equivalent subsampled samples
}
\examples{
df_map <- readMappingData("/home/vicky/Documents/CHMI/Resistome-paper/resistomeAnalysis/db/MAPPING_DATA/nonsubsampled_merged.csv", without_US_duplicates = TRUE)
df_map_sub_saliva <- readMappingData("/home/vicky/Documents/CHMI/Resistome-paper/resistomeAnalysis/db/MAPPING_DATA/subsampled_saliva_merged.csv", without_US_duplicates = TRUE)

df_map_pb_comb <- combineSubsampled(df_map, df_map_sub_saliva, "Drug.Class", "saliva")

}
