% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmaps.R
\name{createRPKMHeatmap}
\alias{createRPKMHeatmap}
\title{Draws a heatmap from a dataframe of ARG abundance for every sample}
\usage{
createRPKMHeatmap(df_map, Cohort, col_vector, show_column_names = FALSE)
}
\arguments{
\item{df_map}{A dataframe of combined non-subsampled or subsampled mapping data and metadata}

\item{Cohort}{A string of the study cohort e.g. "Twin"}

\item{col_vector}{A character vector of colours, named by sample type}

\item{show_column_names}{Logical: If TRUE, shows column names on heatmap}
}
\value{
None
}
\description{
Draws a heatmap from a dataframe of ARG abundance for every sample
}
\examples{
library(RColorBrewer)

df_map <- readMappingData("/home/vicky/Documents/CHMI/Resistome-paper/resistomeAnalysis/db/MAPPING_DATA/nonsubsampled_merged.csv", without_US_duplicates = TRUE)
top_col <- brewer.pal(4, "Paired")
col_vector <- c("stool" = top_col[1], "saliva" = top_col[3])
createRPKMHeatmap(df_map, "Twin", col_vector, show_column_names = TRUE)

}
