% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendrogram.R
\name{createUSDendrogram}
\alias{createUSDendrogram}
\title{Draws a dendrogram of hierachical clustering of ARG rpkm for longitudinal US samples}
\usage{
createUSDendrogram(df_map_dup, coloured_labels)
}
\arguments{
\item{df_map_dup}{A dataframe of combined non-subsampled or subsampled mapping data and metadata, including longitudinal samples}

\item{coloured_labels}{A list of colours named by sample type}
}
\value{
None
}
\description{
Draws a dendrogram of hierachical clustering of ARG rpkm for longitudinal US samples
}
\examples{
library(RColorBrewer)
df_map_dup <- readMappingData("/home/vicky/Documents/CHMI/Resistome-paper/resistomeAnalysis/db/MAPPING_DATA/nonsubsampled_merged.csv", without_US_duplicates = FALSE)
top_col <- rev(brewer.pal(8, "Spectral"))
coloured_labels <- c("stool" = top_col[1], "dorsum of tongue" = top_col[2], "buccal mucosa" = top_col[3], "dental" = top_col[4])
createUSDendrogram(df_map_dup, coloured_labels)

}
