% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_resistome.R
\name{drawCoreARGs}
\alias{drawCoreARGs}
\title{Draw a circular graph showing the core ARGs for each ARG Class}
\usage{
drawCoreARGs(data, hjust_val, bar_label_size, group_label_size, B = 100)
}
\arguments{
\item{data}{A dataframe of the mapping data}

\item{hjust_val}{A numerical vector with the position of the ARG class text for each ARG class}
}
\description{
Draw a circular graph showing the core ARGs for each ARG Class
}
\examples{

}
