% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlationAnalysis.R
\name{drawCorrelationHeatmap}
\alias{drawCorrelationHeatmap}
\title{Draws a heatmap from a dataframe of correlated ARGs and taxa}
\usage{
drawCorrelationHeatmap(cor_df, bottom_margin = 0, left_margin = 0,
  phyla)
}
\arguments{
\item{cor_df}{A dataframe containing correlated ARGs and taxa, rho and adjusted p-values}

\item{bottom_margin}{A numeric value in millimeters for height of column names}

\item{phyla}{A character vector of phyla names for labelling columns of heatmap}

\item{padding}{A numeric value in millimeters for width for row names}
}
\value{
None
}
\description{
Draws a heatmap from a dataframe of correlated ARGs and taxa
}
\examples{
df_map <- readMappingData("/home/vicky/Documents/CHMI/Resistome-paper/resistomeAnalysis/db/MAPPING_DATA/nonsubsampled_merged.csv", without_US_duplicates = TRUE)
metaphlan <- read.csv("/home/vicky/Documents/CHMI/Resistome-paper/resistomeAnalysis/db/METAPHLAN/metaphlan.csv", stringsAsFactors = FALSE, row.names = 1)
metaphlan_rpkm <- combineMetaphlanandARG(df_map, metaphlan)
high_cor_uk_oral <- getSpearmanCorrelation(metaphlan_rpkm, ids = unique(df_map$ID[df_map$Cohort == "Twin" & df_map$sample_type == "saliva"]), taxon_level = "t", taxon_ignore = "@@")
phyla <- c("Firmicutes", "Actinobacteria", "Proteobacteria", "Bacteroidetes", "Candidatus_Saccharibacteria", "Fusobacteria", "Spirochaetes", "Verrucomicrobia", "Ascomycota", "Synergistetes")
h <- drawCorrelationHeatmap(high_cor_uk_oral, 100, 5, phyla)

}
