% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processFiles.R
\name{filterBlastOutput}
\alias{filterBlastOutput}
\title{Filters blast output file}
\usage{
filterBlastOutput(filename, column_names, e_value_filter = NA,
  identity_filter = NA, min_length_filter = NA)
}
\arguments{
\item{filename}{A string of the file name}

\item{column_names}{A character vector of the headers of the blast file}

\item{e_value_filter}{A numeric e-value cut-off}

\item{identity_filter}{A numeric identity_filter cut-off}

\item{min_length_filter}{A numeric minimum contig length cut-off}
}
\description{
Filters blast output file
}
\examples{

}
