% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resistanceMapUse.R
\name{g_legend}
\alias{g_legend}
\title{Extract legend from ggplot object}
\usage{
g_legend(a.gplot)
}
\arguments{
\item{a.gplot}{A ggplot object}
}
\value{
A ggplot object legend
}
\description{
Extract legend from ggplot object
}
