% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processFiles.R
\name{getBestBlastHit}
\alias{getBestBlastHit}
\title{Gets the best BLAST hit from overlapping ORFs}
\usage{
getBestBlastHit(blast_result, qseqid_id)
}
\arguments{
\item{blast_result}{A dataframe of filtered blast results}

\item{qseqid_id}{A character vector of the query sequence IDs to get the best BLAST hit}
}
\description{
Gets the best BLAST hit from overlapping ORFs
}
\examples{

}
