% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProportion.R
\name{getProportion}
\alias{getProportion}
\title{Creates a dataframe that shows the proportion of samples that contain a unique property for every unique property for each country
i.e. the proportion of samples that contain an ARG class for all ARG classes found in column Drug.Class for each country}
\usage{
getProportion(df_map, level)
}
\arguments{
\item{df_map}{A dataframe of combined non-subsampled or subsampled mapping data and metadata}

\item{level}{A string of a column name e.g. Drug.Class}
}
\value{
A dataframe of proportions of samples that contain a unique property for each country
}
\description{
Creates a dataframe that shows the proportion of samples that contain a unique property for every unique property for each country
i.e. the proportion of samples that contain an ARG class for all ARG classes found in column Drug.Class for each country
}
