% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProportion.R
\name{joinProportionAndBootstrap}
\alias{joinProportionAndBootstrap}
\title{Merges dataframes of proportion of samples containing a certain property and confidence intervals of proportions from bootstrapping samples}
\usage{
joinProportionAndBootstrap(df_map, level = "Drug.Class",
  sample_type = "saliva", B = 100)
}
\arguments{
\item{df_map}{A dataframe of combined non-subsampled or subsampled mapping data and metadata}

\item{level}{A string of a column name e.g. "Drug.Class"}

\item{sample_type}{A string of a sample type e.g. "saliva"}

\item{B}{A numerical value of the number of types to sample in bootstrapping}
}
\value{
A dataframe of proportions and 95% confidence intervals of samples that contain a unique property for each Location
}
\description{
Merges dataframes of proportion of samples containing a certain property and confidence intervals of proportions from bootstrapping samples
}
\examples{
df_map <- readMappingData("/home/vicky/Documents/CHMI/Resistome-paper/resistomeAnalysis/db/MAPPING_DATA/nonsubsampled_merged.csv", without_US_duplicates = TRUE)
df_map_pb <- joinProportionAndBootstrap(df_map, "Drug.Class", "saliva")

}
