% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resistanceMapUse.R
\name{mergeData}
\alias{mergeData}
\title{Merge dataframes proportion and confidence intervals of samples containing ARG class and resistanceMap use data}
\usage{
mergeData(percentages_data_list, resistance_map_data)
}
\arguments{
\item{percentages_data_list}{A list of dataframes of proportions containing ARG classes and confidence intervals for each sample type}

\item{resistance_map_data}{A dataframe containing resistanceMap use data}
}
\value{
A dataframe of proportions containing ARG classes and confidence intervals, and resistanceMap DDD per 1000
}
\description{
Merge dataframes proportion and confidence intervals of samples containing ARG class and resistanceMap use data
}
\examples{
df_map <- readMappingData("/home/vicky/Documents/CHMI/Resistome-paper/resistomeAnalysis/db/MAPPING_DATA/nonsubsampled_merged.csv", without_US_duplicates = TRUE)
df_map_pb_saliva_class <- joinProportionAndBootstrap(df_map, "Drug.Class", "saliva")
df_map_pb_dental_class <- joinProportionAndBootstrap(df_map, "Drug.Class", "dental")
rm_use_card <- readResistanceMapUseData("/home/vicky/Documents/CHMI/Resistome-paper/resistomeAnalysis/db/ResistanceMap/resistanceMap_use.csv")

data_merge <- mergeData(list(saliva = df_map_pb_saliva_class, dental = df_map_pb_dental_class), rm_use_card)

}
