% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackedBarChartsRPKM.R
\name{plotARGClassAbundance}
\alias{plotARGClassAbundance}
\title{Plot stacked bar charts of the relative abundance of total rpkm per country, sample type and drug class}
\usage{
plotARGClassAbundance(df_map_abundance, cols)
}
\arguments{
\item{df_map_abundance}{A dataframe of relative abundances created from \code{\link{getRelativeAbundance}}}

\item{cols}{A character vector, named by ARG class}
}
\value{
A list of ggplot object where each graph represents a country
}
\description{
Plot stacked bar charts of the relative abundance of total rpkm per country, sample type and drug class
}
\examples{
library(RColorBrewer)
df_map <- readMappingData("/home/vicky/Documents/CHMI/Resistome-paper/resistomeAnalysis/db/MAPPING_DATA/nonsubsampled_merged.csv", without_US_duplicates = TRUE)
df_map_rel <- getRelativeAbundance(df_map)
cols <- brewer.pal(length(unique(df_map_rel$Drug.Class.Alt)), "Set3")
names(cols) <- unique(df_map_rel$Drug.Class.Alt)
g <- plotARGClassAbundance(df_map_rel[df_map_rel$Country == "HMP1",], cols)

}
