% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARGPhageAnalysis.R
\name{plotARGPhage}
\alias{plotARGPhage}
\title{Plot bar charts of Phage-ARG combinations}
\usage{
plotARGPhage(phage_args_comb, phages_args_pairs, sample_type, Location,
  colours)
}
\arguments{
\item{phage_args_comb}{A dataframe containing the percentage of samples containing each phage and ARG}

\item{phages_args_pairs}{A dataframe containing the percentage of samples containing phage-ARG pairs}

\item{sample_type}{A string of the sample type i.e. saliva}

\item{Location}{A string of the location i.e. "China"}

\item{colours}{A character vector of colours named by drug class}
}
\description{
Plot bar charts of Phage-ARG combinations
}
\examples{

}
