% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackedBarChartsRPKM.R
\name{plotIndividualAbundance}
\alias{plotIndividualAbundance}
\title{Plot stacked bar charts of the relative abundance of total rpkm per individual, sample type and drug class}
\usage{
plotIndividualAbundance(df_map_abundance, cols)
}
\arguments{
\item{df_map_abundance}{A dataframe of relative abundances created from \code{\link{getRelativeAbundanceIndividuals}}}

\item{cols}{A character vector, named by ARG class}
}
\value{
A list of ggplot object where each graph represents a Location
}
\description{
Plot stacked bar charts of the relative abundance of total rpkm per individual, sample type and drug class
}
\examples{
library(RColorBrewer)
df_map <- readMappingData("/home/vicky/Documents/CHMI/Resistome-paper/resistomeAnalysis/db/MAPPING_DATA/nonsubsampled_merged.csv", without_US_duplicates = TRUE)
df_map_rel_ind <- getRelativeAbundanceIndividuals(df_map)
cols <- brewer.pal(length(unique(df_map_rel_ind$Drug.Class.Alt)), "Set3")
names(cols) <- unique(df_map_rel_ind$Drug.Class.Alt)
g <- plotIndividualAbundance(df_map_rel_ind[df_map_rel_ind$Location == "HMP1",], cols)

}
