% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplotsARGRichness.R
\name{plotMultipleARGRichnessGraphs}
\alias{plotMultipleARGRichnessGraphs}
\title{Plots multiple box and whisker diagrams of the ARG richness comparison between paired sample types for every Location}
\usage{
plotMultipleARGRichnessGraphs(ttest_groups, df_map_subsampled_pairs)
}
\arguments{
\item{ttest_groups}{A dataframe of p-values, asterisks, countries and groups.}

\item{df_map_subsampled_pairs}{A dataframe of paired samples containing group column.}
}
\value{
A list of box and whisker ggplot objects.
}
\description{
Plots multiple box and whisker diagrams of the ARG richness comparison between paired sample types for every Location
}
\examples{
df_map_subsampled <- readMappingData("/home/vicky/Documents/CHMI/Resistome-paper/resistomeAnalysis/db/MAPPING_DATA/subsampled_argrich_merged.csv", without_US_duplicates = TRUE)
pair_list <- list(c("stool", "dental"), c("stool", "saliva"), c("dental", "saliva"), c("stool", "dorsum of tongue"), c("stool", "buccal mucosa"), c("dorsum of tongue", "buccal mucosa"), c("dorsum of tongue", "dental"), c("buccal mucosa", "dental"))
df_map_subsampled_pairs <- createPairedData(df_map_subsampled, pair_list)
ttest_groups <- runTtest(df_map_subsampled_pairs)

g <- plotMultipleARGRichnessGraphs(ttest_groups, df_map_subsampled_pairs)

}
