% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplotsRPKM.R
\name{plotMultipleRPKM}
\alias{plotMultipleRPKM}
\title{Plots multiple box and whisker diagrams of total RPKM between paired sample types for every Location}
\usage{
plotMultipleRPKM(df_map_pairs)
}
\arguments{
\item{df_map_pairs}{A dataframe of paired sample types with a column named group i.e. saliva vs. stool}
}
\value{
A list of box and whisker ggplot objects.
}
\description{
Plots multiple box and whisker diagrams of total RPKM between paired sample types for every Location
}
\examples{
df_map <- readMappingData("/home/vicky/Documents/CHMI/Resistome-paper/resistomeAnalysis/db/MAPPING_DATA/nonsubsampled_merged.csv", without_US_duplicates = TRUE)
pair_list <- list(c("stool", "dental"), c("stool", "saliva"), c("dental", "saliva"), c("stool", "dorsum of tongue"), c("stool", "buccal mucosa"), c("dorsum of tongue", "buccal mucosa"), c("dorsum of tongue", "dental"), c("buccal mucosa", "dental"))
df_map_pairs <- createPairedData(df_map, pair_list)

g <- plotMultipleRPKM(df_map_pairs)

}
