% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProportion.R
\name{plotPercentages}
\alias{plotPercentages}
\title{Plots a bar chart of the porportions of samples that contain a property for each Location}
\usage{
plotPercentages(df_map_pb, colours)
}
\arguments{
\item{df_map_pb}{A dataframe of proportions and 95\% confidence intervals of samples that contain a unique property for each Location}

\item{colours}{A character vector of colours, named by Location}
}
\value{
None
}
\description{
Plots a bar chart of the porportions of samples that contain a property for each Location
}
\examples{
library(RColorBrewer)
library(ggplot2)
df_map <- readMappingData("/home/vicky/Documents/CHMI/Resistome-paper/resistomeAnalysis/db/MAPPING_DATA/nonsubsampled_merged.csv", without_US_duplicates = TRUE)
cols <- c("grey20", "grey50", "grey80", "white")
names(cols) <- c("China", "Philippines", "UK", "US")
df_map_pb_saliva_class <- joinProportionAndBootstrap(df_map, "Drug.Class", "saliva")

plotPercentages(df_map_pb_saliva_class, cols) + xlab("ARG class") + ylab("\% saliva samples containing ARG class")

}
