% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplotsRPKM.R
\name{plotRPKM}
\alias{plotRPKM}
\title{Plots a box and whisker diagram between paired samples types}
\usage{
plotRPKM(df_map_pairs_group)
}
\arguments{
\item{df_map_pairs_group}{A dataframe representing one country with paired samples}
}
\value{
A ggplot object
}
\description{
Plots a box and whisker diagram between paired samples types
}
