% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resistotypes.R
\name{plotSampleTypes}
\alias{plotSampleTypes}
\title{Plot principal coordinates with sample type labels}
\usage{
plotSampleTypes(mds)
}
\arguments{
\item{mds}{A dataframe of multidimensional scaled coordinates}
}
\value{
None
}
\description{
Plot principal coordinates with sample type labels
}
\examples{
df_map <- readMappingData("/home/vicky/Documents/CHMI/Resistome-paper/resistomeAnalysis/db/MAPPING_DATA/nonsubsampled_merged.csv", without_US_duplicates = TRUE)
mds <- runPrincipleCoordinateAnalysis(df_map)
plotSampleTypes(mds)

}
