% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProportion.R
\name{plotScatter}
\alias{plotScatter}
\title{Plots a scatter plots of the porportions of samples that contain a property for each Location for subsampled and non-subsampled samples}
\usage{
plotScatter(df_map_pb_comb)
}
\arguments{
\item{df_map_pb_comb}{A dataframe with proportions and confidence intervals from non-subsampled and equivalent subsampled samples}
}
\value{
None
}
\description{
Plots a scatter plots of the porportions of samples that contain a property for each Location for subsampled and non-subsampled samples
}
\examples{
df_map <- readMappingData("/home/vicky/Documents/CHMI/Resistome-paper/resistomeAnalysis/db/MAPPING_DATA/nonsubsampled_merged.csv", without_US_duplicates = TRUE)
df_map_sub_saliva <- readMappingData("/home/vicky/Documents/CHMI/Resistome-paper/resistomeAnalysis/db/MAPPING_DATA/subsampled_saliva_merged.csv", without_US_duplicates = TRUE)
df_map_pb_comb <- combineSubsampled(df_map, df_map_sub_saliva, "Drug.Class", "saliva")

plotScatter(df_map_pb_comb)

}
