% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differentialAnalysis.R
\name{plotVolcano}
\alias{plotVolcano}
\title{Plot Volcano plots of differential analysis}
\usage{
plotVolcano(Country_results, df_map, sample_comp, title)
}
\arguments{
\item{Country_results}{A dataframe of DESeq2 results for one country}

\item{df_map}{A dataframe of combined non-subsampled or subsampled mapping data and metadata}

\item{sample_comp}{A string of sample type comparison e.g. "saliva vs. stool"}

\item{title}{A string of the country name for title of volcano plot}
}
\value{
A ggplot2 object of a volcano plot
}
\description{
Plot Volcano plots of differential analysis
}
\examples{
df_map <- readMappingData("/home/vicky/Documents/CHMI/Resistome-paper/resistomeAnalysis/db/MAPPING_DATA/nonsubsampled_merged.csv", without_US_duplicates = TRUE)
output_list <- runDESeq2(df_map, "US", c("stool", "dorsum of tongue"))
g <- plotVolcano(output_list$results, df_map, "stool vs. saliva", "China")

}
