% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resistanceMapUse.R
\name{readResistanceMapUseData}
\alias{readResistanceMapUseData}
\title{Read resistanceMap use data}
\usage{
readResistanceMapUseData(filename)
}
\arguments{
\item{filename}{A string of the csv filename of resistanceMap use data}
}
\value{
None
}
\description{
Read resistanceMap use data
}
\examples{
rm_use_card <- readResistanceMapUseData("/home/vicky/Documents/CHMI/Resistome-paper/resistomeAnalysis/db/ResistanceMap/resistanceMap_use.csv")

}
