% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resistotypes.R
\name{runPrincipleCoordinateAnalysis}
\alias{runPrincipleCoordinateAnalysis}
\title{Run principal coordinate analysis on mapping data and metadata}
\usage{
runPrincipleCoordinateAnalysis(df_map)
}
\arguments{
\item{df_map}{A dataframe of combined non-subsampled or subsampled mapping data and metadata}
}
\value{
A dataframe of multidimensional scaled coordinates
}
\description{
Run principal coordinate analysis on mapping data and metadata
}
\examples{
df_map <- readMappingData("/home/vicky/Documents/CHMI/Resistome-paper/resistomeAnalysis/db/MAPPING_DATA/nonsubsampled_merged.csv", without_US_duplicates = TRUE)
mds <- runPrincipleCoordinateAnalysis(df_map)

}
