% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplotsARGRichness.R
\name{runTtest}
\alias{runTtest}
\title{Runs a t-test between groups from a dataframe of paired samples}
\usage{
runTtest(df_map_subsampled_pairs)
}
\arguments{
\item{df_map_subsampled_pairs}{A dataframe of paired samples containing a column named "group" i.e. saliva vs. stool}
}
\value{
A dataframe of p-values, asterisks, countries and groups.
}
\description{
Runs a t-test between groups from a dataframe of paired samples
}
\examples{
df_map_subsampled <- readMappingData("/home/vicky/Documents/CHMI/Resistome-paper/resistomeAnalysis/db/MAPPING_DATA/subsampled_argrich_merged.csv", without_US_duplicates = TRUE)
pair_list <- list(c("stool", "dental"), c("stool", "saliva"), c("dental", "saliva"), c("stool", "dorsum of tongue"), c("stool", "buccal mucosa"), c("dorsum of tongue", "buccal mucosa"), c("dorsum of tongue", "dental"), c("buccal mucosa", "dental"))
df_map_subsampled_pairs <- createPairedData(df_map_subsampled, pair_list)

ttest_groups <- runTtest(df_map_subsampled_pairs)

}
