% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processFiles.R
\name{combineBedtools}
\alias{combineBedtools}
\title{Combines bedtools mapping data with metadata and CARD data}
\usage{
combineBedtools(filenames, output_file, metadata_file, card_dir,
  subsampled = FALSE)
}
\arguments{
\item{filenames}{A character vector of bedtools filenames}

\item{output_file}{A string of the location of combined bedtools and metadata file}

\item{metadata_file}{A string of the location of the metadata file}

\item{card_dir}{A string of the CARD metadata directory that includes aro_index and aro_categories_index}

\item{subsampled}{A boolean stating whether reads were subsampled or not before mapping to CARD}
}
\description{
Combines bedtools mapping data with metadata and CARD data
}
\examples{

}
