% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_functions.R
\name{combineMetaphlanSamples}
\alias{combineMetaphlanSamples}
\title{Combine the Metaphlan relative abundances for all samples}
\usage{
combineMetaphlanSamples(filenames, output_file)
}
\arguments{
\item{filenames}{A character vector of filenames of relative abundance output data from Metaphlan}

\item{output_file}{A string of the file path of the output csv file}
}
\value{
None
}
\description{
Combine the Metaphlan relative abundances for all samples
}
