% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differentialAnalysis.R
\name{plotVolcano}
\alias{plotVolcano}
\title{Plot Volcano plots of differential analysis}
\usage{
plotVolcano(Location_results, df_map, title, class_colours, label_size)
}
\arguments{
\item{Location_results}{A dataframe of DESeq2 results for one Location}

\item{df_map}{A dataframe of combined non-subsampled or subsampled mapping data and metadata}

\item{title}{A string of the Location name for title of volcano plot}
}
\value{
A ggplot2 object of a volcano plot
}
\description{
Plot Volcano plots of differential analysis
}
\examples{
df_map <- readMappingData("/home/vicky/Documents/CHMI/Resistome-paper/resistomeAnalysis/db/MAPPING_DATA/nonsubsampled_merged.csv", without_US_duplicates = TRUE)
output_list <- runDESeq2(df_map, "US", c("stool", "dorsum of tongue"))
g <- plotVolcano(output_list$results, df_map, "stool vs. saliva", "China")

}
