% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMappingData.R
\name{readMappingData}
\alias{readMappingData}
\title{Read a csv file of combined mapping data and metadata}
\usage{
readMappingData(filename, without_US_duplicates = TRUE)
}
\arguments{
\item{filename}{A character vector of the csv}

\item{without_US_duplicates}{Logical: If TRUE, includes the US longitudinal data}
}
\value{
A dataframe of processed, combined mapping data and metadata
}
\description{
Read a csv file of combined mapping data and metadata
}
\examples{
df_map <- readMappingData("/home/vicky/Documents/CHMI/Resistome-paper/resistomeAnalysis/db/MAPPING_DATA/nonsubsampled_merged.csv", without_US_duplicates = TRUE)
df_map_dup <- readMappingData("/home/vicky/Documents/CHMI/Resistome-paper/resistomeAnalysis/db/MAPPING_DATA/nonsubsampled_merged.csv", without_US_duplicates = FALSE)

}
